feature: Delete image on image change in edit product

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
-const { validationResult } = require('express-validator')
+const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -132,6 +133,7 @@ exports.postEditProduct = (req, res, next) => {
         return res.redirect('/')
       }
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         return product.update({ title, price, description, imageUrl: image.path })
           .then(result => {
             console.log('Updated Product')
@@ -152,10 +154,16 @@ exports.postEditProduct = (req, res, next) => {
 }
 
 exports.postDeleteProduct = (req, res, next) => {
-  const { productId } = req.body
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+  const { productId } = req.body;
+  Product.findById(productId).then(product => {
+    if (!product) {
+      return next(new Error('Product not found.'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: productId, userId: req.user._id })
+  })
     .then(() => {
-      res.redirect('/admin/products')
+      res.redirect('/admin/products');
     })
     .catch(err => {
       const error = new Error(err);
